+++ assets/javascripts/kitten/index.js
@@ -1,3 +1,6 @@
+// COMPONENTS
+// ----------
+
 // Accessibility
 export { VisuallyHidden } from './components/accessibility/visually-hidden'
 
@@ -185,6 +188,11 @@ export { Overlay } from './components/modals/overlay'
 
 // Navigation
 export { default as HeaderNav } from './components/navigation/header-nav'
+export {
+  MOBILE_HEADER_HEIGHT,
+  TABLET_HEADER_HEIGHT,
+  DESKTOP_HEADER_HEIGHT,
+} from './components/navigation/header-nav/config'
 export { HorizontalNav } from './components/navigation/horizontal-nav'
 export { Pagination } from './components/navigation/pagination'
 export { TabBar } from './components/navigation/tab-bar'
@@ -264,3 +272,33 @@ export { Uploader } from './components/uploaders/uploader'
 export { Video } from './components/videos/video'
 export { GifVideo } from './components/videos/gif-video'
 export { EmbedPlayer } from './components/videos/embed-player'
+
+// CONSTANTS
+// ---------
+
+export { default as COLORS } from './constants/colors-config'
+export {
+  NUM_COLUMNS,
+  CONTAINER_MAX_WIDTH,
+  GUTTER,
+  CONTAINER_PADDING_THIN,
+  CONTAINER_PADDING,
+} from './constants/grid-config'
+export { ScreenConfig } from './constants/screen-config'
+export { default as TYPOGRAPHY } from './constants/typography-config'
+
+// HELPERS
+// -------
+
+export { cssPropertyDistributor } from './helpers/dom/css-property-distributor'
+export { default as domElementHelper } from './helpers/dom/element-helper'
+export {
+  getReactElementsByType,
+  getReactElementsWithoutType,
+} from './helpers/react/react-elements'
+export { default as useContinuousIntersectionObserver } from './helpers/utils/continuous-intersection-hook'
+export { default as useIntersectionObserver } from './helpers/utils/intersection-hook'
+export { default as useLazyObserver } from './helpers/utils/lazy-hook'
+export { default as margin } from './helpers/utils/margin'
+export { pxToRem, stepToRem } from './helpers/utils/typography'
+export { usePrevious } from './helpers/utils/use-previous-hook'