+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Fix `DropdownSelect` item line height and alignment.
+- Fix: `ProgressBar` wasn’t passing class names.
 
 ## [2.104.1] - 2020-12-14