+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix: `VerticalStepper` styles.
 - Fix: Fix className and refacto.
+- Fix: Fix `DropdownSelect` item line height and alignment.
+- Fix: `ProgressBar` wasn’t passing class names.
 
 ## [2.104.1] - 2020-12-14