+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `VerticalStepper` styles.
+- Fix: Fix className and refacto.
 - Fix: Fix `DropdownSelect` item line height and alignment.
 - Fix: `ProgressBar` wasn’t passing class names.